feat: Add list.vue and get-reason-list.ts
This commit adds the `list.vue` file containing a template for displaying a list of reasons. It also includes the `get-reason-list.ts` file which implements an API endpoint to retrieve the list of reasons from a file. The `list.vue` template dynamically fetches the list using the `getReasonList()` function and populates the `ol` element with the reasons. -Jetbrains AI :D